+++ libnavigation-base/src/main/java/com/mapbox/navigation/base/trip/model/roadobject/custom/Custom.kt
@@ -16,8 +16,9 @@ class Custom internal constructor(
     length: Double?,
     location: RoadObjectLocation,
     @RoadObjectProvider.Type provider: String,
+    isUrban: Boolean?,
     nativeRoadObject: com.mapbox.navigator.RoadObject,
-) : RoadObject(id, RoadObjectType.CUSTOM, length, location, provider, nativeRoadObject) {
+) : RoadObject(id, RoadObjectType.CUSTOM, length, location, provider, isUrban, nativeRoadObject) {
 
     /**
      * Returns a string representation of the object.