+++ libnavigation-base/src/main/java/com/mapbox/navigation/base/trip/model/roadobject/tunnel/Tunnel.kt
@@ -18,8 +18,9 @@ class Tunnel internal constructor(
     length: Double?,
     location: RoadObjectLocation,
     @RoadObjectProvider.Type provider: String,
+    isUrban: Boolean?,
     nativeRoadObject: com.mapbox.navigator.RoadObject,
-) : RoadObject(id, RoadObjectType.TUNNEL, length, location, provider, nativeRoadObject) {
+) : RoadObject(id, RoadObjectType.TUNNEL, length, location, provider, isUrban, nativeRoadObject) {
 
     /**
      * Indicates whether some other object is "equal to" this one.